move Attachment component to dedicated folder (prepare for splitup)
